+++ client/src/components/AboutDialog.tsx
@@ -27,7 +27,7 @@ const AboutDialog = (props: AboutDialogProps) => (
 				<h2>
 					<TeamOtpLogo />
 				</h2>
-				<p>Version 0.0.3</p>
+				<p>Version 0.0.4</p>
 				<a href="https://github.com/serkonda7/teamotp" target="_blank" rel="noreferrer">
 					Source code on GitHub
 				</a>